+++ sdk/src/main/java/io/opentdf/platform/sdk/KASClient.java
@@ -20,7 +20,7 @@ import io.opentdf.platform.kas.RewrapResponse;
 import io.opentdf.platform.sdk.Config.KASInfo;
 import io.opentdf.platform.sdk.nanotdf.ECKeyPair;
 import io.opentdf.platform.sdk.nanotdf.NanoTDFType;
-import io.opentdf.platform.sdk.TDF.KasBadRequestException;
+import io.opentdf.platform.sdk.SDK.KasBadRequestException;
 
 import okhttp3.OkHttpClient;
