+++ sdk/src/main/java/io/opentdf/platform/sdk/Manifest.java
@@ -22,7 +22,7 @@ import com.nimbusds.jose.crypto.RSASSASigner;
 import com.nimbusds.jose.crypto.RSASSAVerifier;
 import com.nimbusds.jwt.JWTClaimsSet;
 import com.nimbusds.jwt.SignedJWT;
-import io.opentdf.platform.sdk.TDF.AssertionException;
+import io.opentdf.platform.sdk.SDK.AssertionException;
 import org.apache.commons.codec.binary.Hex;
 import org.erdtman.jcs.JsonCanonicalizer;
