+++ sdk/src/main/java/io/opentdf/platform/sdk/NanoTDF.java
@@ -3,7 +3,7 @@ package io.opentdf.platform.sdk;
 import com.connectrpc.ResponseMessageKt;
 import io.opentdf.platform.policy.kasregistry.ListKeyAccessServersRequest;
 import io.opentdf.platform.policy.kasregistry.ListKeyAccessServersResponse;
-import io.opentdf.platform.sdk.TDF.KasAllowlistException;
+import io.opentdf.platform.sdk.SDK.KasAllowlistException;
 import io.opentdf.platform.sdk.nanotdf.*;
 
 import java.io.IOException;
